feat: Merging to new context
